Changed structure of Trie + TrieNode classes.
Added integer indexing, Integers are now the index into the hashmap and
the sorting element of the treenodes. This is to allow lookup of the
trie from the dictionary tupples, because without an integer lookup we
cannot locate the previous nodes with the information given from the
tuple.

ie, all trie nodes need 3 variables, integer index, char value, and
treemap children.